+++ src/Microsoft.DotNet.Cli.Utils/CommandResolutionStrategy.cs
@@ -1,0 +1,17 @@
+﻿namespace Microsoft.DotNet.Cli.Utils
+{
+    internal enum CommandResolutionStrategy
+    {
+        //command loaded from a nuget package
+        NugetPackage,
+
+        //command loaded from the same directory as the executing assembly
+        BaseDirectory,
+
+        //command loaded from path
+        Path,
+
+        //command not found
+        None
+    }
+}
